Renamed array_to_vpack to vpack_conversion
